+++ examples/exporter-statsd-example/src/main.rs
@@ -15,6 +15,8 @@ struct Recorder;
 
 // In order to use this with `axum_prometheus`, we must implement `MakeDefaultHandle`.
 impl MakeDefaultHandle for Recorder {
+    // We don't need to return anything meaningful from here (unlike PrometheusHandle)
+    // Let's just return an empty tuple.
     type Out = ();
 
     fn make_default_handle() -> Self::Out {
@@ -26,9 +28,6 @@ impl MakeDefaultHandle for Recorder {
             .expect("Could not create StatsdRecorder");
 
         metrics::set_boxed_recorder(Box::new(recorder)).unwrap();
-        // We don't need to return anything meaningful from here (unlike PrometheusHandle)
-        // Let's just return an empty tuple.
-        ()
     }
 }
 
@@ -37,7 +36,7 @@ async fn main() {
     tracing_subscriber::registry()
         .with(
             tracing_subscriber::EnvFilter::try_from_default_env()
-                .unwrap_or_else(|_| "exporter-statsd-example=debug".into()),
+                .unwrap_or_else(|_| "exporter_statsd_example=debug".into()),
         )
         .with(tracing_subscriber::fmt::layer())
         .init();