+++ src/lifecycle/body.rs
@@ -1,5 +1,6 @@
-use super::{Callbacks, FailedAt};
+use super::{Callbacks, FailedAt, OnBodyChunk};
 use futures_core::ready;
+use http::HeaderValue;
 use http_body::Body;
 use pin_project::pin_project;
 use std::{
@@ -11,18 +12,24 @@ use tower_http::classify::ClassifyEos;
 
 /// Response body for [`LifeCycle`].
 #[pin_project]
-pub struct ResponseBody<B, C, Callbacks, CallbacksData> {
+pub struct ResponseBody<B, C, Callbacks, OnBodyChunk, CallbacksData> {
     #[pin]
     pub(super) inner: B,
-    pub(super) parts: Option<(C, Callbacks, CallbacksData)>,
+    pub(super) parts: Option<(C, Callbacks)>,
+    pub(super) callbacks_data: CallbacksData,
+    pub(super) on_body_chunk: OnBodyChunk,
+    pub(super) content_length: Option<HeaderValue>,
 }
 
-impl<B, C, CallbacksT, CallbacksData> Body for ResponseBody<B, C, CallbacksT, CallbacksData>
+impl<B, C, CallbacksT, OnBodyChunkT, CallbacksData> Body
+    for ResponseBody<B, C, CallbacksT, OnBodyChunkT, CallbacksData>
 where
     B: Body,
     B::Error: fmt::Display + 'static,
     C: ClassifyEos,
     CallbacksT: Callbacks<C::FailureClass, Data = CallbacksData>,
+    OnBodyChunkT: OnBodyChunk<B::Data, Data = CallbacksData>,
+    CallbacksData: Clone,
 {
     type Data = B::Data;
     type Error = B::Error;
@@ -33,21 +40,29 @@ where
     ) -> Poll<Option<Result<Self::Data, Self::Error>>> {
         let this = self.project();
 
-        let result = ready!(this.inner.poll_data(cx));
+        let body_size = this.inner.size_hint().exact();
+        let Some(result) = ready!(this.inner.poll_data(cx)) else {
+            return Poll::Ready(None);
+        };
+
+        let body_size = body_size.or_else(|| {
+            this.content_length
+                .as_ref()
+                .and_then(|cl| cl.to_str().ok())
+                .and_then(|cl| cl.parse().ok())
+        });
 
         match result {
-            None => Poll::Ready(None),
-            Some(Ok(chunk)) => {
-                if let Some((_, callbacks, callbacks_data)) = &this.parts {
-                    callbacks.on_body_chunk(&chunk, callbacks_data);
-                }
+            Ok(chunk) => {
+                this.on_body_chunk
+                    .call(&chunk, body_size, this.callbacks_data);
 
                 Poll::Ready(Some(Ok(chunk)))
             }
-            Some(Err(err)) => {
-                if let Some((classify_eos, callbacks, callbacks_data)) = this.parts.take() {
+            Err(err) => {
+                if let Some((classify_eos, callbacks)) = this.parts.take() {
                     let classification = classify_eos.classify_error(&err);
-                    callbacks.on_failure(FailedAt::Body, classification, callbacks_data);
+                    callbacks.on_failure(FailedAt::Body, classification, this.callbacks_data);
                 }
 
                 Poll::Ready(Some(Err(err)))
@@ -65,18 +80,18 @@ where
 
         match result {
             Ok(trailers) => {
-                if let Some((classify_eos, callbacks, callbacks_data)) = this.parts.take() {
+                if let Some((classify_eos, callbacks)) = this.parts.take() {
                     let trailers = trailers.as_ref();
                     let classification = classify_eos.classify_eos(trailers);
-                    callbacks.on_eos(trailers, classification, callbacks_data);
+                    callbacks.on_eos(trailers, classification, this.callbacks_data.clone());
                 }
 
                 Poll::Ready(Ok(trailers))
             }
             Err(err) => {
-                if let Some((classify_eos, callbacks, callbacks_data)) = this.parts.take() {
+                if let Some((classify_eos, callbacks)) = this.parts.take() {
                     let classification = classify_eos.classify_error(&err);
-                    callbacks.on_failure(FailedAt::Trailers, classification, callbacks_data);
+                    callbacks.on_failure(FailedAt::Trailers, classification, this.callbacks_data);
                 }
 
                 Poll::Ready(Err(err))